refactor: move reference_syndrome from backend/ to api/
reference_syndrome orchestrates the full pipeline (strip noise via
optimizer, lower via backend, run one shot via svm), so living in
backend/ with includes pointing at optimizer and svm misrepresented its
role. Move to a new clifft/api/ directory that sits above
backend/optimizer/svm and owns this kind of full-pipeline glue.
