+++ src/python/bindings.cc
@@ -1,5 +1,5 @@
+#include "clifft/api/reference_syndrome.h"
 #include "clifft/backend/backend.h"
-#include "clifft/backend/reference_syndrome.h"
 #include "clifft/circuit/circuit.h"
 #include "clifft/circuit/parser.h"
 #include "clifft/frontend/frontend.h"
